Makefile.container: run clean before world
On my local GBP instance, when a build fails, I have an email sent to
me. This email contains attachments for `emerge --info`, the Jenkins
"console" log, as well as a tar.gz of the entire contents of
/var/tmp/portage in the container. What happens is that over time the
/var/tmp/portage directory fills up due to packages getting updated and
not rebuilt after the initial failure. So this tar file gets bigger and
bigger until eventually my email server regects the email because the
message is too big.

This change cleans the tmpdir before every world update.